+++ task_manager/static/scripts/routes.js
