+++ task_manager/static/scripts/templates.js
