+++ task_manager/static/scripts/views/home.js
