+++ task_manager/static/scripts/views/taskHandlers.js
