+++ task_manager/static/scripts/views/viewUtils.js
