+++ task_manager/templates/index_static.html
